Removing ActionBar is now a user preference.
The user, if they are running SDK >= 11 will have an option in
SettingsActivity to enable or disable the ActionBar for ConsoleActivity.
The option exists because some Honeycomb and Ice Cream Sandwich users
might need the ActionBar to access the Activity menu, while others may
prefer to not see the ActionBar at all since it takes up a fair amount
of space.

Users SDK <= 10 will not be given this option and for them
ConsoleActivity will not have a TitleBar, just as it was in previous
versions.